Remove dependency on libretro's include files (as well as functions) and completetly switching to c99 data types without an extra include file.

+++ mednafen/sound/Blip_Buffer.h
@@ -134,7 +134,7 @@ typedef struct
 	int delta_factor;
 } Blip_Synth;
 
-static INLINE void Blip_Synth_set_volume(Blip_Synth* synth, double v, int range)
+static INLINE void Blip_Synth_set_volume(Blip_Synth* synth, float v, int range)
 {
 	synth->delta_factor = ((v * (1.0 / (range < 0 ? -range : range))) * (1L << blip_sample_bits) + 0.5);
 }
